Extract header and subheader from CatalogTree component, fix minor styling issues
