+++ airbyte-webapp/src/views/Connection/CatalogTree/CatalogSection.module.scss
@@ -1,7 +1,7 @@
 @use "../../../scss/colors";
 
 .streamFieldTableContainer {
-  margin-left: 83px;
+  margin-left: 85px;
   background: colors.$grey-50;
 }
