+++ airbyte-webapp/src/views/Connection/CatalogTree/CatalogTreeRows.module.scss
@@ -1,0 +1,9 @@
+.container {
+  margin-bottom: 29px;
+  max-height: 600px;
+  overflow-y: auto;
+
+  --webkit-overlay: true;
+
+  width: 100%;
+}
