+++ airbyte-webapp/src/views/Connection/CatalogTree/CatalogTreeSubheader.module.scss
@@ -1,0 +1,19 @@
+%catalogHeader {
+  min-height: 33px;
+  margin-bottom: 0 !important;
+  margin-left: 20px;
+}
+
+.catalogSubheader {
+  @extend %catalogHeader;
+
+  padding-top: 10px;
+  margin-left: 155px;
+}
+
+.readonlyCatalogSubheader {
+  @extend %catalogHeader;
+
+  padding-top: 10px;
+  margin-left: 115px;
+}
